Separate file for pure_colors.py
