+++ colorpy/ciexyz.py
@@ -881,102 +881,6 @@ def xyz_from_spectrum (spectrum_array):
     return xyz
 
 #
-# Miscellaneous stuff for spectral line colors that does not really belong here.
-#
-
-def get_normalized_spectral_line_colors (
-    brightness = 1.0,
-    num_purples = 0,
-    dwl_angstroms = 10):
-    '''Get an array of xyz colors covering the visible spectrum.
-    Optionally add a number of 'purples', which are colors interpolated between the color
-    of the lowest wavelength (violet) and the highest (red).
-
-    brightness - Desired maximum rgb component of each color.  Default 1.0.  (Maxiumum displayable brightness)
-    num_purples - Number of colors to interpolate in the 'purple' range.  Default 0.  (No purples)
-    dwl_angstroms - Wavelength separation, in angstroms (0.1 nm).  Default 10 A. (1 nm spacing)
-    '''
-    # get range of wavelengths, in angstroms, so that we can have finer resolution than 1 nm
-    wl_angstrom_range = range (10*start_wl_nm, 10*(end_wl_nm + 1), dwl_angstroms)
-    # get total point count
-    num_spectral = len (wl_angstrom_range)
-    num_points   = num_spectral + num_purples
-    xyzs = numpy.empty ((num_points, 3))
-    # build list of normalized color x,y values proceeding along each wavelength
-    i = 0
-    for wl_A in wl_angstrom_range:
-        wl_nm = wl_A * 0.1
-        xyz = xyz_from_wavelength (wl_nm)
-        colormodels.xyz_normalize (xyz)
-        xyzs [i] = xyz
-        i += 1
-    # interpolate from end point to start point (filling in the purples)
-    first_xyz = xyzs [0]
-    last_xyz  = xyzs [num_spectral - 1]
-    for ipurple in range (0, num_purples):
-        t = float (ipurple) / float (num_purples - 1)
-        omt = 1.0 - t
-        xyz = t * first_xyz + omt * last_xyz
-        colormodels.xyz_normalize (xyz)
-        xyzs [i] = xyz
-        i += 1
-    # scale each color to have the max rgb component equal to the desired brightness
-    for i in range (0, num_points):
-        rgb = colormodels.brightest_rgb_from_xyz (xyzs [i], brightness)
-        xyzs [i] = colormodels.xyz_from_rgb (rgb)
-    # done
-    return xyzs
-
-def get_normalized_spectral_line_colors_annotated (
-    brightness = 1.0,
-    num_purples = 0,
-    dwl_angstroms = 10):
-    '''Get an array of xyz colors covering the visible spectrum.
-    Optionally add a number of 'purples', which are colors interpolated between the color
-    of the lowest wavelength (violet) and the highest (red).
-    A text string describing the color is supplied for each color.
-
-    brightness - Desired maximum rgb component of each color.  Default 1.0.  (Maxiumum displayable brightness)
-    num_purples - Number of colors to interpolate in the 'purple' range.  Default 0.  (No purples)
-    dwl_angstroms - Wavelength separation, in angstroms (0.1 nm).  Default 10 A. (1 nm spacing)
-    '''
-    # get range of wavelengths, in angstroms, so that we can have finer resolution than 1 nm
-    wl_angstrom_range = range (10*start_wl_nm, 10*(end_wl_nm + 1), dwl_angstroms)
-    # get total point count
-    num_spectral = len (wl_angstrom_range)
-    num_points   = num_spectral + num_purples
-    xyzs = numpy.empty ((num_points, 3))
-    names = []
-    # build list of normalized color x,y values proceeding along each wavelength
-    i = 0
-    for wl_A in wl_angstrom_range:
-        wl_nm = wl_A * 0.1
-        xyz = xyz_from_wavelength (wl_nm)
-        colormodels.xyz_normalize (xyz)
-        xyzs [i] = xyz
-        name = '%.1f nm' % wl_nm
-        names.append (name)
-        i += 1
-    # interpolate from end point to start point (filling in the purples)
-    first_xyz = xyzs [0]
-    last_xyz  = xyzs [num_spectral - 1]
-    for ipurple in range (0, num_purples):
-        t = float (ipurple) / float (num_purples - 1)
-        omt = 1.0 - t
-        xyz = t * first_xyz + omt * last_xyz
-        colormodels.xyz_normalize (xyz)
-        xyzs [i] = xyz
-        name = '%03d purple' % math.floor (1000.0 * t + 0.5)
-        names.append (name)
-        i += 1
-    # scale each color to have the max rgb component equal to the desired brightness
-    for i in range (0, num_points):
-        rgb = colormodels.brightest_rgb_from_xyz (xyzs [i], brightness)
-        xyzs [i] = colormodels.xyz_from_rgb (rgb)
-    # done
-    return (xyzs, names)
-
-#
 # Main.
 #
