+++ colorpy/misc.py
@@ -27,15 +27,6 @@ universe_patch_plot () -
     This originally caused some controversy when the (correct) xyz color was incorrectly reported as light green.
     The authors also consider several other white points, here we just use the default (normally D65).
 
-spectral_colors_patch_plot () -
-    Colors of the pure spectral lines.
-
-spectral_colors_plus_purples_patch_plot () -
-    Colors of the pure spectral lines plus purples.
-
-perceptually_uniform_spectral_colors () -
-    Patch plot of (nearly) perceptually equally spaced colors, covering the pure spectral lines plus purples.
-
 License:
 
 Copyright (C) 2008 Mark Kness
@@ -59,11 +50,7 @@ along with ColorPy.  If not, see <http://www.gnu.org/licenses/>.
 '''
 from __future__ import print_function
 
-import math
-import numpy
-
 import colormodels
-import ciexyz
 import plots
 
 # Some sample lists of displayable RGB colors as hex strings
@@ -139,7 +126,9 @@ def colorstring_patch_plot (colorstrings, color_names, title, filename, num_acro
         filename,
         num_across=num_across)
 
-# Patch plots from xyz color values
+#
+# Some example patch plots for xyz color values.
+#
 
 def MacBeth_ColorChecker_patch_plot ():
     ''' MacBeth ColorChecker Chart. '''
@@ -222,100 +211,6 @@ def universe_patch_plot ():
         'Average Color of the Universe\nhttp://www.pha.jhu.edu/~kgb/cosspec/',
         'Universe')
 
-# Pure spectral colors
-
-def spectral_colors_patch_plot ():
-    '''Colors of the pure spectral lines.'''
-    xyzs = ciexyz.get_normalized_spectral_line_colors (brightness=1.0, num_purples=0, dwl_angstroms=10)
-    plots.xyz_patch_plot (
-        xyzs, None, 'Colors of pure spectral lines', 'Spectral', num_across=20)
-
-
-def spectral_colors_plus_purples_patch_plot ():
-    '''Colors of the pure spectral lines plus purples.'''
-    xyzs = ciexyz.get_normalized_spectral_line_colors (brightness=1.0, num_purples=200, dwl_angstroms=10)
-    plots.xyz_patch_plot (
-        xyzs, None, 'Colors of pure spectral lines plus purples', 'SpectralPlusPurples', num_across=20)
-
-# An attempt to get a perceptually equally spaced (almost) subset of the pure spectral colors
-
-def perceptually_uniform_spectral_colors (
-    brightness = 1.0,
-    plot_name  = 'PerceptuallyEqualColors',
-    plot_title = 'Perceptually (almost) Equally Spaced Pure Colors',
-    table_name = 'percep_equal_names.txt'):
-    '''Patch plot of (nearly) perceptually equally spaced colors, covering the pure spectral lines plus purples.'''
-    # TODO - This may or may not be quite right...
-    # get pure colors
-#    xyzs = ciexyz.get_normalized_spectral_line_colors (brightness=1.0, num_purples=200, dwl_angstroms=1)
-    (xyzs, names) = ciexyz.get_normalized_spectral_line_colors_annotated (brightness=brightness, num_purples=200, dwl_angstroms=1)
-    (num_colors, num_columns) = xyzs.shape
-
-    # pick these two functions for either Luv or Lab
-    uniform_from_xyz = colormodels.luv_from_xyz
-    xyz_from_uniform = colormodels.xyz_from_luv
-    #uniform_from_xyz = colormodels.lab_from_xyz
-    #xyz_from_uniform = colormodels.xyz_from_lab
-
-    # convert colors to a nearly perceptually uniform space
-    uniforms = numpy.empty ((num_colors, 3))
-    for i in range (0, num_colors):
-        uniforms [i] = uniform_from_xyz (xyzs [i])
-    # determine spacing
-    sum_ds = 0.0
-    dss = numpy.empty ((num_colors, 1))
-    for i in range (0, num_colors-1):
-        dri = uniforms [i+1] - uniforms [i]
-        dsi = math.sqrt (numpy.dot (dri, dri))
-        dss [i] = dsi
-        sum_ds += dsi
-    # last point closes the curve
-    dri = uniforms [0] - uniforms [num_colors - 1]
-    dsi = math.sqrt (numpy.dot (dri, dri))
-    dss [num_colors - 1] = dsi
-    sum_ds += dsi
-    # pick out subsamples as evenly spaced as possible
-    num_samples = 160
-    ds_avg = sum_ds / float (num_samples - 1)
-    E_indices = []
-    index = 0
-    count = 0
-    need = 0.0
-    while True:
-        while need > 1.0e-10:
-            need -= dss [index]
-            index += 1
-        E_indices.append (index)
-        need += ds_avg
-        count += 1
-        if count >= num_samples:
-            break
-    # patch plot and save names
-    xyz_list = []
-    fil = open (table_name, 'wt')
-    fil.write ('%s\n' % plot_title)
-    fil.write ('Name iRGB\n')
-    fil.write ('\n')
-    for index in E_indices:
-        uniform_color = uniforms [index]
-        uniform_xyz   = xyz_from_uniform (uniform_color)
-        uniform_irgb  = colormodels.irgb_from_xyz (uniform_xyz)
-        uniform_name  = names [index]
-        xyz_list.append (uniform_xyz)
-        fil.write ('%s %s\n' % (uniform_name, str (uniform_irgb)))
-    fil.close ()
-    plots.xyz_patch_plot (
-        xyz_list, None, plot_title, plot_name, num_across=20)
-
-def perceptually_uniform_spectral_color_plots ():
-    brightness_list = [1.0, 0.9, 0.8, 0.75, 0.6, 0.5, 0.4, 0.3, 0.25]
-    for brightness in brightness_list:
-        ibright = math.floor (100.0 * brightness + 0.5)
-        plot_name  = 'PerceptuallyEqualColors_%d' % ibright
-        plot_title = 'Perceptually (almost) Equally Spaced Pure Colors %d%%' % ibright
-        table_name = 'percep_equal_names_%d.txt' % ibright
-        perceptually_uniform_spectral_colors (brightness, plot_name, plot_title, table_name)
-
 #
 # Main.
 #
@@ -327,14 +222,10 @@ def figures ():
     colorstring_patch_plot (hsv_colors, None, 'HSV Colormap', 'hsv')
     colorstring_patch_plot (jet_colors, None, 'Jet Colormap', 'jet')
     colorstring_patch_plot (primary_colors, primary_names, 'Primary Colors', 'primary', num_across=4)
-    # patch charts of xyz color tables
+    # Example patch plots of xyz colors.
     MacBeth_ColorChecker_patch_plot ()
     chemical_solutions_patch_plot ()
     universe_patch_plot ()
-    # pure colors
-    spectral_colors_patch_plot ()
-    spectral_colors_plus_purples_patch_plot ()
-    perceptually_uniform_spectral_color_plots ()
 
 
 if __name__ == '__main__':
