+++ colorpy/test_colormodels.py
@@ -29,6 +29,7 @@ import unittest
 
 import colormodels
 import ciexyz
+import pure_colors
 
 # Functions to calculate the cutoff point between various algorithms.
 # These do not really belong here...
@@ -193,7 +194,7 @@ class TestColormodels(unittest.TestCase):
     def test_clipping(self, verbose=False):
         ''' Test the various color clipping methods. '''
         # This is just a coverage test.
-        xyz_colors = ciexyz.get_normalized_spectral_line_colors ()
+        xyz_colors = pure_colors.get_normalized_spectral_line_colors ()
         num_wl = xyz_colors.shape[0]
         for i in range (num_wl):
             # Get rgb values for standard add white clipping.
